backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
joblib==1.3.2
python-multipart==0.0.9
pydantic==2.6.1
pydantic[email]==2.6.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
motor==3.1.1
pymongo==4.3.3
python-dotenv==1.0.1
torch==2.2.1
torchvision==0.17.1
pillow==10.2.0
requests==2.31.0
bcrypt==4.1.2
kaggle==1.5.16
python-jose[cryptography]==3.3.0
fastapi-cors==0.0.6
tqdm==4.66.1